requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.5.0 